WebView for Collection Author Added Colors, Font Extensions Moved
