+++ iOS-FakeNFT-Extended/DesignSystem/Fonts.swift
@@ -1,4 +1,5 @@
 import UIKit
+import SwiftUI
 
 extension UIFont {
     // Ниже приведены примеры шрифтов, настоящие шрифты надо взять из фигмы
@@ -16,4 +17,13 @@ extension UIFont {
     // Caption Fonts
     static var caption1 = UIFont.systemFont(ofSize: 15, weight: .regular)
     static var caption2 = UIFont.systemFont(ofSize: 13, weight: .regular)
+}
+
+
+extension Font {
+    static var bodyBold = Font.system(size: 17, weight: .bold)
+    static var body22Bold = Font.system(size: 22, weight: .bold)
+    static var medium10 = Font.system(size: 10, weight: .regular)
+    static var regular13 = Font.system(size: 13, weight: .regular)
+    static var regular15 = Font.system(size: 15, weight: .regular)
 }
