+++ iOS-FakeNFT-Extended/UI/Catalog/CatalogView.swift
@@ -18,18 +18,6 @@ struct CatalogView: View {
     }
 }
 
-extension Font {
-    static var bodyBold = Font.system(size: 17, weight: .bold)
-    static var body22Bold = Font.system(size: 22, weight: .bold)
-    static var medium10 = Font.system(size: 10, weight: .regular)
-    static var regular13 = Font.system(size: 13, weight: .regular)
-    static var regular15 = Font.system(size: 15, weight: .regular)
-}
-
-extension Color {
-    static let blueUniversal = Color(red: 10.0 / 255.0, green: 132.0 / 255.0, blue: 255.0 / 255.0)
-}
-
 #Preview {
     CatalogView()
 }
