+++ iOS-FakeNFT-Extended/UI/WebView.swift
@@ -1,0 +1,21 @@
+//
+//  WebView.swift
+//  iOS-FakeNFT-Extended
+//
+//  Created by Vadzim on 11.02.26.
+//
+
+import SwiftUI
+import WebKit
+
+struct WebView: UIViewRepresentable {
+    let url: URL
+
+    func makeUIView(context: Context) -> WKWebView {
+        WKWebView()
+    }
+
+    func updateUIView(_ uiView: WKWebView, context: Context) {
+        uiView.load(URLRequest(url: url))
+    }
+}
